Move standard function to own class
