+++ src/namespace.js
@@ -1,6 +1,6 @@
 import fetch from 'node-fetch';
 import { getGeoServerResponseText, GeoServerResponseError } from './util/geoserver.js';
-import GeoServerRestClient from '../geoserver-rest-client.js'
+import AboutClient from './about.js'
 
 /**
  * Client for GeoServer namespace
@@ -91,7 +91,7 @@ export default class NamespaceClient {
    *
    * @throws Error if request fails
    *
-   * @returns {Object} An object describing the namespace
+   * @returns {Object} An object describing the namespace or undefined if it cannot be found
    */
   async get (name) {
     const auth =
@@ -104,7 +104,7 @@ export default class NamespaceClient {
       }
     });
     if (!response.ok) {
-      const grc = new GeoServerRestClient(this.url, this.user, this.password);
+      const grc = new AboutClient(this.url, this.user, this.password);
       if (await grc.exists()) {
         // GeoServer exists, but requested item does not exist,  we return empty
         return;
