+++ test/test.js
@@ -20,12 +20,12 @@ const geoServerVersion = process.env.GEOSERVER_VERSION;
 
 describe('Basic GeoServer', () => {
   it('should exist', async () => {
-    const result = await grc.exists();
+    const result = await grc.about.exists();
     expect(result).to.be.true;
   });
 
   it('returns correct version', async () => {
-    const result = await grc.getVersion();
+    const result = await grc.about.getVersion();
     expect(result.about.resource[0].Version).to.equal(geoServerVersion);
   })
 });
